+++ src/lib/display/VacancyCard_home.svelte
